Removed helper functions - added the two used functions inside the app - removed the rest
