+++ apps/dash-brain-connectivity/app.py
@@ -9,17 +9,14 @@ import dash_bootstrap_components as dbc
 import plotly.express as px
 import nibabel as nib
 import numpy as np
-from nilearn import datasets, image
+from nilearn import image
 import pandas as pd
 import dash_table
+from plotly import graph_objects as go
 
 from nibabel.affines import apply_affine
 from matplotlib import pyplot as plt
 
-import sys
-
-sys.path.append("./")
-import helper_functions as hf
 
 stylesheets = [dbc.themes.BOOTSTRAP]
 app = dash.Dash(__name__, external_stylesheets=stylesheets)
@@ -44,6 +41,48 @@ def get_atlas(atlas_key):
     atlas_img = image.resample_to_img(tmp_img, template_img, interpolation="nearest")
     parcellation = atlas_img.get_fdata().squeeze().astype(np.uint8)
     return atlas_table, atlas_description, atlas_reference, atlas_connmat, parcellation
+
+
+def make_heatmap(df, colormap=None):
+    axis_ticks = df.columns
+    n_elements = len(axis_ticks)
+    gap_width = 20 / n_elements
+    if n_elements > 100:
+        gap_width = 0
+    layout = go.Layout(
+        yaxis=dict(
+            autorange="reversed",
+            scaleanchor="x",
+            constrain="domain",
+            scaleratio=1,
+            showgrid=True,
+            showticklabels=False,
+            tickvals=np.arange(-0.5, df.shape[0], 1),
+        ),
+        xaxis=dict(
+            constrain="domain",
+            tickvals=np.arange(-0.5, df.shape[0], 1),
+            showgrid=True,
+            showticklabels=False,
+        ),
+        margin=dict(l=0, r=0, b=0, t=0, pad=0),
+    )
+    heatmap = go.Heatmap(
+        z=df,
+        zmin=0,
+        zmax=1,
+        x=axis_ticks,
+        y=axis_ticks,
+        xgap=gap_width,
+        ygap=gap_width,
+        hoverongaps=False,
+        autocolorscale=True if colormap is None else False,
+        colorscale=colormap,
+    )
+    fig = go.Figure(heatmap, layout)
+    fig.update_traces(showscale=True)
+
+    return fig
 
 
 # Load the initial data
@@ -143,18 +182,27 @@ navbar = dbc.Navbar(
 
 # Layout Slicer Positions
 saggital_slicer_card = dbc.Card(
-    [dbc.CardHeader("Saggital View"), dbc.CardBody(dbc.CardBody(slicer_list[0])),],
+    [
+        dbc.CardHeader("Saggital View"),
+        dbc.CardBody(dbc.CardBody(slicer_list[0])),
+    ],
     color="dark",
     inverse=True,
 )
 coronal_slicer_card = dbc.Card(
-    [dbc.CardHeader("Coronal View"), dbc.CardBody(slicer_list[1]),],
+    [
+        dbc.CardHeader("Coronal View"),
+        dbc.CardBody(slicer_list[1]),
+    ],
     color="dark",
     inverse=True,
 )
 
 axial_slicer_card = dbc.Card(
-    [dbc.CardHeader("Axial View"), dbc.CardBody(dbc.CardBody(slicer_list[2])),],
+    [
+        dbc.CardHeader("Axial View"),
+        dbc.CardBody(dbc.CardBody(slicer_list[2])),
+    ],
     color="dark",
     inverse=True,
 )
@@ -241,7 +289,10 @@ connectivity_info = html.Div(
         dbc.Row(
             [
                 dbc.Col("The current region: ", width=3),
-                dbc.Col(html.Pre(id="position-region-info"), width="auto",),
+                dbc.Col(
+                    html.Pre(id="position-region-info"),
+                    width="auto",
+                ),
             ],
         ),
         dbc.Row(
@@ -253,7 +304,10 @@ connectivity_info = html.Div(
         dbc.Row(
             [
                 dbc.Col("with a strength of: ", width=3),
-                dbc.Col(html.Pre(id="connection-conn-info"), width="auto",),
+                dbc.Col(
+                    html.Pre(id="connection-conn-info"),
+                    width="auto",
+                ),
             ],
         ),
     ],
@@ -375,7 +429,11 @@ table_card = dbc.Card(
 connectivity_card = dbc.Card(
     [
         dbc.CardHeader("Functional Connectivity"),
-        dbc.CardBody([dcc.Graph(id="conn_mat"),]),
+        dbc.CardBody(
+            [
+                dcc.Graph(id="conn_mat"),
+            ]
+        ),
         dbc.CardFooter(
             [
                 html.Div(
@@ -523,10 +581,17 @@ def position_info(seed_region, brain_region, region_names, raw_conn_mat):
     Input("store_graph_region", "data"),
     Input("store_conn_thr_mat", "data"),
     Input("store_parcellation", "data"),
-    [State("region_table", "data"), State("threshold-mode", "value"),],
+    [
+        State("region_table", "data"),
+        State("threshold-mode", "value"),
+    ],
 )
 def update_brain_overlay(
-    graph_region, conn_mat_data, parcellation_data, table_data, threshold,
+    graph_region,
+    conn_mat_data,
+    parcellation_data,
+    table_data,
+    threshold,
 ):
     conn_mat = np.array(conn_mat_data)
     parcellation = np.array(parcellation_data).astype(np.uint8)
@@ -740,6 +805,22 @@ def connectivity_heatmap(graph_region, slicer_region, conn_mat, region_names):
             line=dict(color="#b8daff", width=4),
             orientation="v" if graph_region is not None else "h",
         )
+    return fig
+
+
+def add_region_shape(fig, region_id, line=None, orientation="h"):
+    # We just use the position along the y axis here but because the matrix is symmetrical, there is no difference
+    draw_pos = np.where(np.array(fig.data[0]["y"]) == region_id)[0][0]
+    fig.add_shape(
+        type="rect",
+        xref="x",
+        yref="y",
+        x0=-0.5 if orientation == "h" else draw_pos - 0.5,
+        y0=draw_pos - 0.5 if orientation == "h" else -0.5,
+        x1=len(fig.data[0]["x"]) - 0.5 if orientation == "h" else draw_pos + 0.5,
+        y1=draw_pos + 0.5 if orientation == "h" else len(fig.data[0]["y"]) - 0.5,
+        line=line if line is not None else dict(color="LightSeaGreen", width=3),
+    )
     return fig
