--- apps/dash-brain-connectivity/helper_functions/__init__.py
@@ -1,7 +1,0 @@
-from .connectivity import (
-    compute_connectivity_matrix,
-    make_circos,
-    make_heatmap,
-    compute_region_table,
-    add_region_shape,
-)
